function_app/requirements.txt
azure-functions>=1.20.0
azure-identity>=1.16.0
azure-storage-file-datalake>=12.14.0
azure-storage-blob>=12.20.0
